refactor mcp dto braced normalizer
